# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/confirm/{confirm_uid}": {
            "get": {
                "description": "Подтверждает регистрацию пользователя",
                "tags": [
                    "auth"
                ],
                "summary": "Подтверждение регистрации",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uid конфирмации",
                        "name": "confirm_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auth/new_password": {
            "post": {
                "description": "Меняет пароль пользователя на новый",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Восстановление пароля",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task-manager-backend_internal_app_api.ChangePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Обновляет JWT по refresh токену\nДля того что бы обновить токен надо быть\nаунтифицированным",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновить JWT",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task-manager-backend_internal_app_api.Refresh"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task-manager-backend_internal_app_api.Tokens"
                        }
                    }
                }
            }
        },
        "/auth/restore_password": {
            "post": {
                "description": "Отправляет ссылку на страницу с восстановлением пароля",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Отправка ссылки для восстановления пароля",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task-manager-backend_internal_app_api.RestorePasswordEmail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "description": "Вход в систему по логину и хешу-пароля",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход в систему",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task-manager-backend_internal_app_api.Auth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task-manager-backend_internal_app_api.Tokens"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Прямая регистрация нового пользователя в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task-manager-backend_internal_app_api.Auth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_app_api.Auth": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "pwd_hash": {
                    "type": "string"
                }
            }
        },
        "internal_app_api.ChangePassword": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "restore_uid": {
                    "type": "string"
                }
            }
        },
        "internal_app_api.Refresh": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "internal_app_api.RestorePasswordEmail": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "internal_app_api.Tokens": {
            "type": "object",
            "properties": {
                "tokens": {
                    "$ref": "#/definitions/users.Session"
                }
            }
        },
        "task-manager-backend_internal_app_api.Auth": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "pwd_hash": {
                    "type": "string"
                }
            }
        },
        "task-manager-backend_internal_app_api.ChangePassword": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "restore_uid": {
                    "type": "string"
                }
            }
        },
        "task-manager-backend_internal_app_api.Refresh": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "task-manager-backend_internal_app_api.RestorePasswordEmail": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "task-manager-backend_internal_app_api.Tokens": {
            "type": "object",
            "properties": {
                "tokens": {
                    "$ref": "#/definitions/users.Session"
                }
            }
        },
        "users.Session": {
            "type": "object",
            "properties": {
                "refresh": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}